Infer pointer-backed structures from fixed offsets

+++ decomp/ST.exe/functions/004C6C70/decomp.c
@@ -9,7 +9,7 @@ FUN_004c6c70(void *this,undefined4 param_1,undefined4 param_2,undefined4 param_3
     thunk_FUN_004c6bb0(this,param_1,param_2,param_3);
   }
   else {
-    iVar1 = thunk_FUN_004ea620((int)this);
+    iVar1 = thunk_FUN_004ea620(this);
     if (iVar1 != 0) {
       *(undefined4 *)((int)this + 0x400) = 0;
       *(undefined4 *)((int)this + 0x3fc) = 1;
