+++ decomp/ST.exe/functions/004CE770/decomp.c
@@ -3,7 +3,7 @@ undefined4 __thiscall FUN_004ce770(void *this,int param_1)
 
 {
   if ((param_1 == 1) && (*(int *)((int)this + 0x4d0) == 0)) {
-    thunk_FUN_004ce7a0((int)this);
+    thunk_FUN_004ce7a0(this);
   }
   return 0;
 }
