+++ decomp/ST.exe/functions/0074ABC4/decomp.c
@@ -11,7 +11,7 @@ undefined4 * __thiscall FUN_0074abc4(void *this,undefined4 *param_1,undefined4 p
   *(undefined ***)((int)this + 0x10) = &PTR_FUN_007a1590;
   *(undefined ***)((int)this + 200) = &PTR_FUN_007a1568;
   *(undefined ***)((int)this + 0xcc) = &PTR_FUN_007a1550;
-  FUN_0074ac6e((int)this);
+  FUN_0074ac6e(this);
   return this;
 }
