+++ decomp/ST.exe/functions/0074B96B/decomp.c
@@ -3,7 +3,7 @@ undefined4 * __thiscall FUN_0074b96b(void *this,undefined4 *param_1)
 
 {
   if (param_1 != this) {
-    FUN_0074bce4((int)this);
+    FUN_0074bce4(this);
     FUN_0074bc89(this,param_1);
   }
   return this;
